refactor: 권한 관련 상수 클래스 위치 변경
