+++ src/main/java/nextstep/subway/auth/application/AuthService.java
@@ -1,6 +1,6 @@
 package nextstep.subway.auth.application;
 
-import static nextstep.subway.auth.common.Constants.*;
+import static nextstep.subway.common.constants.AuthConstants.*;
 
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.auth.dto.TokenRequest;
@@ -9,7 +9,6 @@ import nextstep.subway.auth.infrastructure.JwtTokenProvider;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class AuthService {
