+++ src/main/java/nextstep/subway/common/constants/AuthConstants.java
@@ -1,6 +1,6 @@
-package nextstep.subway.auth.common;
+package nextstep.subway.common.constants;
 
-public final class Constants {
+public final class AuthConstants {
     public static final String INVALID_EMAIL = "이메일을 찾을 수 없습니다.";
     public static final String INVALID_PASSWORD = "패스워드가 일치하지 않습니다";
     public static final String INVALID_TOKEN = "유효 하지 않은 토큰입니다.";
